// libs/admin-plugins/vendor-order/src/generated-admin-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
  Date: any;
};
// ...
export type QueryVendorsArgs = {
  options?: Maybe<VendorListOptions>;
};
// ...
export enum DeletionResult {
  /** The entity was successfully deleted */
  Deleted = 'DELETED',
  /** Deletion did not take place, reason given in message */
  NotDeleted = 'NOT_DELETED'
}
// ...
export type DeletionResponse = {
  __typename?: 'DeletionResponse';
  result: DeletionResult;
  message?: Maybe<Scalars['String']>;
};
// ...
export type CreateCustomerInput = {
  title?: Maybe<Scalars['String']>;
  firstName: Scalars['String'];
  lastName: Scalars['String'];
  phoneNumber?: Maybe<Scalars['String']>;
  emailAddress: Scalars['String'];
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type Success = {
  __typename?: 'Success';
  success: Scalars['Boolean'];
};
// ...
export enum VendorAccountType {
  Current = 'CURRENT',
  Saving = 'SAVING'
}

export type VendorInput = {
  firstName: Scalars['String'];
  lastName: Scalars['String'];
  email: Scalars['String'];
  password?: Maybe<Scalars['String']>;
  phone: Scalars['String'];
  GSTINID: Scalars['String'];
  state: Scalars['String'];
  ownerName: Scalars['String'];
  ownerEmail: Scalars['String'];
};

export type VendorInfoInput = {
  vendorId: Scalars['String'];
  brandName: Scalars['String'];
  regAddress: Scalars['String'];
  postalCode: Scalars['String'];
  city: Scalars['String'];
  countryCode: Scalars['String'];
  panno: Scalars['String'];
  GSTINID: Scalars['String'];
  state: Scalars['String'];
  ADHAR: Scalars['String'];
  aboutUs: Scalars['String'];
  staffEmail: Scalars['String'];
  phone: Scalars['String'];
  currencyCode?: Maybe<Scalars['String']>;
  defaultTaxZoneId?: Maybe<Scalars['Int']>;
  defaultShippingZoneId?: Maybe<Scalars['Int']>;
};

export type VendorBankInput = {
  vendorId: Scalars['String'];
  account: Scalars['String'];
  code: Scalars['String'];
  address: Scalars['String'];
  type: VendorAccountType;
  isCheck?: Maybe<Scalars['String']>;
};

export type VendorContactInput = {
  vendorId: Scalars['String'];
  contactName: Scalars['String'];
  email: Scalars['String'];
  phone: Scalars['String'];
};

export type VendorMarketingContactInput = {
  vendorId: Scalars['String'];
  name: Scalars['String'];
  emailAddress: Scalars['String'];
  phone: Scalars['String'];
};

export type UpdateVendorInput = {
  id: Scalars['ID'];
  firstName: Scalars['String'];
  lastName: Scalars['String'];
  email: Scalars['String'];
  phone: Scalars['String'];
  GSTINID: Scalars['String'];
  state: Scalars['String'];
  ownerName: Scalars['String'];
  ownerEmail: Scalars['String'];
};

export type UpdateVendorInfoInput = {
  id: Scalars['ID'];
  brandName: Scalars['String'];
  regAddress: Scalars['String'];
  postalCode: Scalars['String'];
  city: Scalars['String'];
  countryCode: Scalars['String'];
  panno: Scalars['String'];
  GSTINID: Scalars['String'];
  state: Scalars['String'];
  ADHAR: Scalars['String'];
  aboutUs: Scalars['String'];
  staffEmail: Scalars['String'];
  phone: Scalars['String'];
  currencyCode?: Maybe<Scalars['String']>;
  defaultTaxZoneId?: Maybe<Scalars['Int']>;
  defaultShippingZoneId?: Maybe<Scalars['Int']>;
};

export type UpdateVendorBankInput = {
  id: Scalars['ID'];
  account: Scalars['String'];
  code: Scalars['String'];
  address: Scalars['String'];
  type: VendorAccountType;
  isCheck?: Maybe<Scalars['String']>;
};

export type UpdateVendorContactInput = {
  id: Scalars['ID'];
  contactName: Scalars['String'];
  email: Scalars['String'];
  phone: Scalars['String'];
};

export type UpdateVendorMarketingInput = {
  id: Scalars['ID'];
  name: Scalars['String'];
  emailAddress: Scalars['String'];
  phone: Scalars['String'];
};
// ...
export enum VerifyResult {
  Success = 'SUCCESS',
  Fail = 'FAIL'
}
// ...
export type VendorListOptions = {
  skip?: Maybe<Scalars['Int']>;
  take?: Maybe<Scalars['Int']>;
  sort?: Maybe<VendorSortParameter>;
  filter?: Maybe<VendorFilterParameter>;
};
// ...
export type VendorFilterParameter = {
  firstName?: Maybe<StringOperators>;
  lastName?: Maybe<StringOperators>;
  email?: Maybe<StringOperators>;
  phone?: Maybe<StringOperators>;
  GSTINID?: Maybe<StringOperators>;
  state?: Maybe<StringOperators>;
  ownerName?: Maybe<StringOperators>;
  ownerEmail?: Maybe<StringOperators>;
  verified?: Maybe<BooleanOperators>;
};

export type VendorSortParameter = {
  id?: Maybe<SortOrder>;
  firstName?: Maybe<SortOrder>;
  lastName?: Maybe<SortOrder>;
  email?: Maybe<SortOrder>;
  phone?: Maybe<SortOrder>;
  GSTINID?: Maybe<SortOrder>;
  state?: Maybe<SortOrder>;
  ownerName?: Maybe<SortOrder>;
  ownerEmail?: Maybe<SortOrder>;
};